+++ package.json
@@ -1,10 +1,10 @@
 {
   "name": "zero",
-  "version": "0.1.6",
+  "version": "0.2.0",
   "description": "",
   "main": "src/ZeroBot/shard.js",
   "scripts": {
-    "start": "node src/ZeroBot/shard.js",
+    "start": "clear && node src/ZeroBot/shard.js",
     "server": "node src/ZeroServer/server.js"
   },
   "repository": {
