+++ src/ZeroBot/events/ready.js
@@ -1,3 +1,9 @@
-module.exports = (client) => {
-  console.log(`Zero: Bot Ready! I'm logged in as ${client.user.username}!`);
+const { Events } = require("../structure/Client");
+
+module.exports = {
+  name: Events.ClientReady,
+  once: true,
+  execute(client) {
+    console.log(`Zero: Bot Ready! I'm logged in as ${client.user.username}!`);
+  },
 };
