+++ src/ZeroBot/structure/Client.js
@@ -1,6 +1,9 @@
 const { Client, Collection, GatewayIntentBits, Events } = require("discord.js");
 const { readdirSync } = require("fs");
 
+const path = require("node:path");
+const fs = require("node:fs");
+
 const { prefix, token } = require("../../../config.json");
 
 const PathCmd = "./src/ZeroBot/commands";
@@ -13,7 +16,7 @@ let pull;
 class ZeroClient extends Client {
   constructor(options) {
     super(options);
-    this.OS = process.platform
+    this.OS = process.platform;
     this.commands = new Collection();
     this.aliases = new Collection();
     this.prefix = prefix;
@@ -31,7 +34,7 @@ class ZeroClient extends Client {
 
       for (let file of commands) {
         pull = require(`../commands/${dir}/${file}`);
-        if (pull.name) {        
+        if (pull.name) {
           this.commands.set(pull.name, pull);
         } else {
           continue;
@@ -43,24 +46,19 @@ class ZeroClient extends Client {
   }
 
   LoadEvents() {
-    const events = readdirSync(PathEvent).filter((file) =>
-      file.endsWith(".js")
-    );
-    for (let file of events) {
-      try {
-        pull = require(`../events/${file}`);
-        pull.event = pull.event || file.replace(".js", "");
+    const eventsPath = path.join(__dirname, "../events");
+    const eventFiles = fs
+      .readdirSync(eventsPath)
+      .filter((file) => file.endsWith(".js"));
 
-        if (pull.event == "ready") {
-          this.on(Events.ClientReady, pull.bind(null, this));
-        }
-        if (pull.event == "message") {
-          this.on(Events.MessageCreate, pull.bind(null, this));
-        }
+    for (const file of eventFiles) {
+      const filePath = path.join(eventsPath, file);
+      const event = require(filePath);
 
-        //this.on(pull.event, pull.bind(null, this));
-      } catch (err) {
-        console.log(err);
+      if (event.once) {
+        this.on(event.name, (...args) => event.execute(this, ...args));
+      } else {
+        this.on(event.name, (...args) => event.execute(this, ...args));
       }
     }
   }
@@ -68,5 +66,6 @@ class ZeroClient extends Client {
 
 module.exports = {
   ZeroClient,
-  GatewayIntentBits
+  GatewayIntentBits,
+  Events,
 };
