+++ src/ZeroBot/utils/events.js
@@ -1,0 +1,10 @@
+const { Events } = require("discord.js")
+
+
+let events = {
+        "message": Events.MessageCreate
+}
+
+module.exports = {
+    events
+}
